requirements.txt
pymupdf>=1.23.0
python-docx>=1.1.0
langchain-groq>=0.1.0
python-dotenv>=1.0.0
flask>=3.0.0
werkzeug>=3.0.0
flask
openpyxl
